Found the problem with keyboard input
The problem is known as ghosting and it is a hardware problem.
On my laptop keyboard, I cannot press s, e and d at the same time.
There are other combinations of keys, I cannot press.
It has to do with how the laptop is wired.
We must change the idea to one where you press (toggle) the piano keys
you want to play, and then commit them to the musical staff by pressing Enter