Found the source of the bug
The staffs of sheet music were being rendered twice.
That is because the render function we use was stored in state
using useState, and it was being invoked on the first render.
To fix the issue we store, not the render function, but the
anonymous function that returns the render function.
So instead of

  `setRenderFunction(renderFunction);`

we do

  `setRenderFunction(() => renderFunction)`